+++ packages/storybook/stories/prosess/uttak/Uttak.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 import UttakProsessIndex from '@fpsak-frontend/prosess-uttak/src/UttakProsessIndex';
-import { uttaksplaner, behandlingPersonMap } from '@fpsak-frontend/prosess-uttak/src/components/types/testdata';
+import { uttaksplaner, behandlingPersonMap } from '@fpsak-frontend/prosess-uttak/src/components/dto/testdata';
 
 export default {
   title: 'prosess/prosess-pleiepenger-uttak',
